Add FP16 support in DLTensor.

+++ dali/test/python/test_operator_slice.py
@@ -362,7 +362,7 @@ def test_slice_synth_data_vs_numpy():
                 ((200,400,3), "HWC", None, "HW", types.INT64, types.UINT8),
                 ((200,400,3), "HWC", None, "C", types.FLOAT, None),
                 ((200,400,3), "HWC", (1,0), None, types.FLOAT, types.FLOAT16),
-                ((200,400,3), "HWC", (0,1), None, types.INT32, types.FLOAT16),
+                ((200,400,3), "HWC", (0,1), None, types.FLOAT16, types.FLOAT16),
                 ((200,400,3), "HWC", (2,), None, types.FLOAT, None),
                 ((200,), "H", (0,), None, types.FLOAT, None),
                 ((200,), "H", None, "H", types.FLOAT, None),
